initial separation of controller types
