+++ Inc/controller_const.h
@@ -2,18 +2,16 @@
 // Created by Tun Kapgen on 09.06.20.
 //
 
-#ifndef C_IMPLEMENTATION_CONTROLLER_CONST_H
-#define C_IMPLEMENTATION_CONTROLLER_CONST_H
+#ifndef CONTROLLER_CONST_H
+#define CONTROLLER_CONST_H
 
 
 /* Constants */
 #include "controller.h"
 #include <math.h>
 
-#define AVIONICS 1
-
-#define CONTROL_ACTIVE true //
-#define LQR_ACTIVE true                                                                         //
+/* CONTROLLER_TYPE: 0=TEST_CONTROLLER, 1=LQR, 2=MPC */
+#define CONTROLLER_TYPE 1                                                                    
 #define CONTROLLER_SAMPLING_FREQ 100                                                            // Hz
 #define POLY_DEG 30                                                                             // NEEDS CHANGE
 #define OPT_TRAJ_CONTROL_INPUT 0.495415061477727                                                    // -
@@ -23,10 +21,13 @@
 #define NUM_GAINS NUM_POLYFITS - 1                                                              // -
 #define M_AW 0.005                                                                              // -
 #define DELTA_T 1.0f / CONTROLLER_SAMPLING_FREQ                                                 // s
-#define TARGET_AGOGEE 1435.0                                                                    // m
-#define TEST_CONTROLLER_USE_VELOCITY false                                                       // -
-#define TEST_CONTROLLER_MAX_VELOCITY 3.0f                                                       // m/s
-#define TEST_CONTROLLER_MAX_ALTITUDE 20.0f                                                      // m
+#define TARGET_AGOGEE 1435.0                                                                    // m                                                   // m
+
+#if CONTROLLER_TYPE == 0
+    #define TEST_CONTROLLER_USE_VELOCITY false                                                       // -
+    #define TEST_CONTROLLER_MAX_VELOCITY 3.0f                                                       // m/s
+    #define TEST_CONTROLLER_MAX_ALTITUDE 20.0f   
+#endif
 
 /* Types */
 typedef struct {
@@ -46,8 +47,6 @@ typedef struct {
     double poly_coeff[NUM_POLYFITS][POLY_DEG+1];
 
 } control_data_t;
-
-void init_coeff(control_data_t *control_data);
 
 
 #endif //C_IMPLEMENTATION_CONTROLLER_CONST_H
